Praveen V- added unique user viewbook in F.E & B.E and signup valid v2

+++ angularapp/src/app/components/customerside/mybooking/mybooking.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { BookEvent } from 'src/app/class/bookevent';
+import { User } from 'src/app/class/user';
 import { BookEventService } from 'src/app/services/bookevent.service';
+import { DataService } from 'src/app/services/data.service';
 
 @Component({
   selector: 'app-viewbooking',
@@ -10,25 +12,45 @@ import { BookEventService } from 'src/app/services/bookevent.service';
 })
 export class MybookingComponent implements OnInit{
   // events: bookevent=new bookevent();
-  events: BookEvent[] = [];
-  constructor(private bookEventService: BookEventService, private router: Router) {}
+  events?: BookEvent[];
+  userid= 0;
+  cusId='';
+  id= 0;
+  customer: User = new User();
+  constructor(private bookEventService: BookEventService,
+              private router: Router,private route: ActivatedRoute,
+              private data: DataService) {}
 
   ngOnInit(): void {
-    this.bookEventService.viewEvent().subscribe(data =>{
-      this.events = data;
-  });
-}
+  this.data.share4.subscribe(x => this.cusId = x);
+  console.log(this.cusId);
+  this.userid=Number(this.cusId);
 
-updateEvent(eventId: number): any{
-  this.router.navigate(['user/editEvent',eventId]);
-}
+  this.bookEventService.getcustomerByuserId(this.userid).subscribe(data =>
+    {
+      this.customer = data;
+      this.id  = this.customer.userId;
+      console.log(this.id);
 
-deleteBookedEvent(eventId: number): any{
+      this.bookEventService.getBookingById(this.id).subscribe(data2 => {
+        this.events = data2;
+        console.log(this.events);
+      });
 
-    this.bookEventService.deleteEvent(eventId).subscribe( data =>{
-      this.ngOnInit();
-      console.log('deleted');
-    });
+    }, error => console.log(error)
+    );
+  }
 
-}
+  updateEvent(eventId: number): any{
+    this.router.navigate(['user/editEvent',eventId]);
+  }
+
+  deleteBookedEvent(eventId: number): any{
+    if(window.confirm('You are going to delete the Booked Event !')){
+      this.bookEventService.deleteEvent(eventId).subscribe( data =>{
+        this.ngOnInit();
+        console.log('deleted');
+      });
+    }
+  }
 }
